Move logic for latest version, more error checking
